Budgetory Stock Pulse/requirements.txt
streamlit
yfinance
google-generativeai
feedparser
python-dotenv